# Fix: Improves detection of self-referential functions (fixes #2363)