--- a/lib/rules/no-unused-vars.js
+++ b/lib/rules/no-unused-vars.js
@@ -27,9 +27,14 @@ module.exports = function(context) {
         }
     }
 
+    //--------------------------------------------------------------------------
+    // Helpers
+    //--------------------------------------------------------------------------
+
+
     /**
      * Determines if a given variable is being exported from a module.
-     * @param {Variable} variable EScope variable object.
+     * @param {Variable} variable - EScope variable object.
      * @returns {boolean} True if the variable is exported, false if not.
      * @private
      */
@@ -52,7 +57,7 @@ module.exports = function(context) {
 
     /**
      * Determines if a reference is a read operation.
-     * @param {Reference} ref - an escope Reference
+     * @param {Reference} ref - An escope Reference
      * @returns {Boolean} whether the given reference represents a read operation
      * @private
      */
@@ -61,30 +66,47 @@ module.exports = function(context) {
     }
 
     /**
-     * Determine if an identifier is referencing the enclosing function name.
-     * @param {Reference} ref The reference to check.
+     * Determine if an identifier is referencing an enclosing function name.
+     * @param {Reference} ref - The reference to check.
+     * @param {ASTNode[]} nodes - The candidate function nodes.
      * @returns {boolean} True if it's a self-reference, false if not.
      * @private
      */
-    function isSelfReference(ref) {
+    function isSelfReference(ref, nodes) {
+        var scope = ref.from;
 
-        if (ref.from.type === "function" && ref.from.block.id) {
-            return ref.identifier.name === ref.from.block.id.name;
+        while (scope != null) {
+            if (nodes.indexOf(scope.block) >= 0) {
+                return true;
+            }
+
+            scope = scope.upper;
         }
 
         return false;
     }
 
     /**
-     * Determines if a reference should be counted as a read. A reference should
-     * be counted only if it's a read and it's not a reference to the containing
-     * function declaration name.
-     * @param {Reference} ref The reference to check.
-     * @returns {boolean} True if it's a value read reference, false if not.
-     * @private
+     * Determines if the variable is used.
+     * @param {Variable} variable - The variable to check.
+     * @param {Reference[]} [references=variable.references] - The variable references to check.
+     * @returns {boolean} True if the variable is used
      */
-    function isValidReadRef(ref) {
-        return isReadRef(ref) && !isSelfReference(ref);
+    function isUsedVariable(variable, references) {
+        var functionNodes = variable.defs.filter(function (def) {
+            return def.type === "FunctionName";
+        }).map(function (def) {
+            return def.node;
+        }),
+            isFunctionDefinition = functionNodes.length > 0;
+
+        if (!references) {
+            references = variable.references;
+        }
+
+        return references.some(function (ref) {
+            return isReadRef(ref) && !(isFunctionDefinition && isSelfReference(ref, functionNodes));
+        });
     }
 
     /**
@@ -132,7 +154,7 @@ module.exports = function(context) {
                     continue;
                 }
 
-                if (variables[i].references.filter(isValidReadRef).length === 0 && !isExported(variables[i])) {
+                if (!isUsedVariable(variables[i]) && !isExported(variables[i])) {
                     unused.push(variables[i]);
                 }
             }
@@ -140,6 +162,10 @@ module.exports = function(context) {
 
         return [].concat.apply(unused, scope.childScopes.map(getUnusedLocals));
     }
+
+    //--------------------------------------------------------------------------
+    // Public
+    //--------------------------------------------------------------------------
 
     return {
         "Program:exit": function(programNode) {
@@ -149,15 +175,34 @@ module.exports = function(context) {
 
             // determine unused globals
             if (config.vars === "all") {
-                var unresolvedRefs = globalScope.through.filter(isValidReadRef).map(function(ref) {
-                    return ref.identifier.name;
-                });
+                var ref, name;
+                // Avoid inherited properties that could induce false positives (e.g. "constructor")
+                var unresolvedRefs = Object.create(null);
+
+                // Search for read references, and store them in a dictionary by name
+                for (i = 0, l = globalScope.through.length; i < l; ++i) {
+                    ref = globalScope.through[i];
+                    name = ref.identifier.name;
+
+                    if (isReadRef(ref)) {
+                        if (!unresolvedRefs[name]) {
+                            unresolvedRefs[name] = [];
+                        }
+                        unresolvedRefs[name].push(ref);
+                    }
+                }
 
                 for (i = 0, l = globalScope.variables.length; i < l; ++i) {
-                    if (unresolvedRefs.indexOf(globalScope.variables[i].name) < 0 &&
-                            !globalScope.variables[i].eslintUsed && !isExported(globalScope.variables[i])) {
+                    name = globalScope.variables[i].name;
+
+                    var isUsed = unresolvedRefs[name] &&
+                        isUsedVariable(globalScope.variables[i], unresolvedRefs[name]);
+
+                    if (!isUsed && !globalScope.variables[i].eslintUsed &&
+                            !isExported(globalScope.variables[i])) {
                         unused.push(globalScope.variables[i]);
                     }
+
                 }
             }