This implementation considers inner self-reference as invalid (i.e. not valid reads):

```javascript
// foo is defined but not used
function foo() {
  return function () {
     return foo;
  }
}
```