@nzakas 2015-05-08
I think this looks good. Can you do a perf test? I always get a bit concerned when we start looking up scopes frequently. You can run `npm run perf` before and after your change to see if there's a difference.