@jrvidal 2015-05-08
Before:
```
CPU Speed is 800 with multiplier 7500000
Performance Run #1:  1271.060549ms
Performance Run #2:  1283.5308420000001ms
Performance Run #3:  1272.612908ms
Performance Run #4:  1295.978007ms
Performance Run #5:  1274.169107ms
Performance budget ok:  1274.169107ms (limit: 9375ms)
```

After:
```
CPU Speed is 800 with multiplier 7500000
Performance Run #1:  1258.473082ms
Performance Run #2:  1272.925776ms
Performance Run #3:  1274.682477ms
Performance Run #4:  1258.292188ms
Performance Run #5:  1278.11336ms
Performance budget ok:  1272.925776ms (limit: 9375ms)
```